Endpoints can be APIs version-less
Since Keystone V3, the endpoints don't need API versions numbers to be specified
in their respective URL [1]. Effectively, Keystone automatically adjusts to
either versions depending on the context.

For instance, using the default values will generate the following catalog:
keystone identity RegionOne
  publicURL: http://127.0.0.1:5000
  internalURL: http://127.0.0.1:5000
  adminURL: http://127.0.0.1:35357

Note:
Openstack wiki pages might not reflect this entirely yet and there are still
examples using the version number in the URLs.

This patch preserves the version parameter for the endpoint manifests but also
allows the version to be removed properly when it's blandked (version => '').

[1] https://github.com/openstack/keystone/blob/master/doc/source/http-api.rst

Change-Id: I8584d1c1e0a111d016ee507f8b004d5e131a2700